fix sleep function (tho calculating it preciselly is still pending), and add a forrest gump program that runs forever

+++ Kernel/include/drivers/pitDriver.h
@@ -1,10 +1,10 @@
 #ifndef PIT_DRIVER_H
 #define PIT_DRIVER_H
-
+#include <stdint.h>
 
 // Initializes the PIT, setting the frequency to 
 void init_pit();
 
-void sleep(int milis);
+void sleep(uint64_t milis);
 
 #endif
